Fixes missing reference to the `tokenize` function
When the `make_token` function is changed to return a `Result`, the `tokenize` function also needs to be changed. Adds this explanation to the same section of the article and provides the code. This is added after the introduction of the `?` operator so that the code can use this operator as well.