Fix the signature of `parse_tokens` and tests
The signature of `parse_tokens` didn't match the tests, the function wanted ownership but the tests were only giving a borrow.

In this case, it's a better opportunity to teach the slice concept and use that. Updates the `parse_tokens`, `parse_array`, and `parse_object` functions to take token slices. Updates the tests to use arrays and pass those in as slices.