Fix emoji section
In Safari on iOS, it was trying to actually render the surrogate characters, which just rendered as an orange rectangle since it was just half of the pair. I only intended to write the unicode code itself (the ASCII letters/numbers of it), so changed the formatting to "U+XXXX" which shouldn't actually try to render it.

Reworded and rearranged some of the related paragraphs to make it make more sense.