Fix toggle transition on route navigation
In dark mode, the toggle kept resetting to the left and sliding over to the right on route navigation

Since in Next.js `Link` we're not doing a full page refresh, we can keep the toggle value in state to avoid it changing. Full page refreshes will still show a brief slide over, which isn't as jarring since the whole page is loading anyways.